Completed Header
Completed Header component for use across all pages.

-------- index.css --------
TailwindCSS was used for styling that will work on every page, and necessary segments for tailwind and font styling was added to index.css.

-------- Header.jsx --------
The header currently lacks functionality with localStorage, of which I opted for instead of cookies. Pages should work for base use, but functionality for the Community dropdown's filter idea has not been implemented yet