WIP Styling Header/Home
Working on redoing the header and home page so that it's responsive with mobile.

--------- Header.jsx ---------
Major work in progress, About halfway done with actual code, but I still need to style and make sure functionality is there.

--------- Home.jsx ---------
Also a WIP, Improvements on hero section have been made, but resizing needs to be done for larger devices. The quick links need to also be resized but eventually restyled entirely because the styling right now does not work