Update Sign in and backend functionality
Updated the sign in and backend functionality with new features and some improvements. Multer was added as cloudinary uploads will be necessary for employers adding jobs and users adding profiles. Worked on signing in users, and data pushes to MongoDB but there are issues with the profile icon in the header.